% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosge.R
\name{toSge}
\alias{toSge}
\title{ToSge}
\usage{
toSge(cores = NULL, name = NULL, queue = NULL, log = NULL,
  venv = NULL, script = NULL, memmory = NULL, email = NULL)
}
\arguments{
\item{cores}{Do you agree dogs are the best pet? Defaults to TRUE.}

\item{name}{Name of the job}

\item{queue}{Queue where to run the script. choose one of the follow possibilities: 'imppc', 'imppc12', 'imppcv3'}

\item{log}{Full path of log file, the name of the file will be the date + name of the job}

\item{venv}{Name of python virtual environment if it is necessary}

\item{script}{Full path of script to run}

\item{memmory}{Amount of RAM GB}

\item{email}{Email where to send notifications}
}
\description{
Lazy R parser for R
}
\examples{
cores = '16'
name = 'mapInsu'
queue = 'imppcv3'
log = '/imppc/labs/lplab/share/marc/insulinomas/logs'
script = '/imppc/labs/lplab/share/marc/bin/Strelka.sh'
memmory = '8'
email = 'clusterigtpmsubirana@gmail.com'
toSge(cores=cores,
      name=name,
      queue=queue,
      log=log,
      venv=venv,
      script=script,
      memmory=memmory,
      email=email)

toSge()
}
